chore(format-check): satisfy SwiftLint thresholds without baseline bumps
- `RegistryInvariantTests`: rename short `db` parameter in the GRDB
  read closure to `database` (identifier_name violation).
- `ProfileSession.init`: drop the per-call arguments to `finishInit`
  — they were redundant pass-throughs of stored properties already
  set by the time `finishInit` runs. Reading from `self` keeps the
  initialiser body under SwiftLint's `function_body_length`
  threshold without joining multi-line calls or collapsing blanks
  (per the project's "no cosmetic compensating shrinks" rule).
- `SyncCoordinator`: extract `saveStateSerialization` /
  `deleteStateSerialization` into a new sibling extension file
  `SyncCoordinator+StatePersistence.swift`, freeing 15 lines from
  the main file so the addition of the `sharedRegistryStore`
  property's doc-comment doesn't push it over the 400-line
  `file_length` budget. Fits the existing pattern of `+Lifecycle`,
  `+Backfill`, `+Hooks`, etc.

Validation: `just format-check` clean, `just test` (mac + iOS) — 2706
+ 2687 tests pass.
